+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Update dependency io.cucumber:messages up to v26 ((#38)[https://github.com/cucumber/query/pull/38])
+- Add `timestamp` attribute ((#45)[https://github.com/cucumber/junit-xml-formatter/pull/45])
 
 ## [0.5.0] - 2024-06-22
 ### Added